Reorganized the folder to categorize classes based on their function. Commented code in more low-level classes --working on commenting everything that needs to be explained.

+++ game_engine/src/game_engine/Display.java
@@ -9,9 +9,12 @@ import java.awt.Paint;
 import java.awt.Toolkit;
 import java.awt.image.BufferStrategy;
 
+import abstracts.GameEntity;
+
 public class Display extends Canvas {
 
-	// Various system properties to take advantage of OpenGL and graphics acceleration
+	// Various system properties to take advantage of OpenGL and graphics
+	// acceleration
 	static {
 		System.setProperty("sun.java2d.transaccel", "True");
 		// System.setProperty("sun.java2d.trace", "timestamp,log,count");
@@ -22,8 +25,7 @@ public class Display extends Canvas {
 
 	private static final long serialVersionUID = 1L;
 	private BufferStrategy strategy;
-	private Paint backgroundGradient; // TODO put this in its own class w/ buffer strategy
-
+	private Paint backgroundGradient;
 	public static Dimension SIZE;
 
 	public Display(Dimension dims) {
@@ -32,41 +34,66 @@ public class Display extends Canvas {
 
 	public void setupWindow() {
 
-		// Uses a double buffer, and ignores repaint requests since we are handling the graphics
+		// Uses a double buffer, and ignores repaint requests since we are
+		// handling the graphics
 		// updates manually
 		this.createBufferStrategy(2);
 		strategy = this.getBufferStrategy();
 		this.setIgnoreRepaint(true);
 
 		// Sets the background of the game window
-		backgroundGradient = new GradientPaint(0, 0, Color.gray, SIZE.width, SIZE.height,
-				Color.lightGray.brighter());
+		backgroundGradient = new GradientPaint(0, 0, Color.gray, SIZE.width,
+				SIZE.height, Color.lightGray.brighter());
 
 		// Allows this component to get input from mouse and keyboard
 		setFocusable(true);
 	}
 
+	/**
+	 * This function takes a float used for interpolation of graphics between
+	 * two states to ensure smooth graphics rendering. Then it renders the
+	 * current game state by first filling in the background, then rendering
+	 * each game object. Finally, it renders the GUI.
+	 * 
+	 * @param alpha
+	 *            the interpolation value
+	 */
 	public void render(double alpha) {
+		// Gets the graphics strategy and sets the background to the one defined
+		// by backgroundGradient
 		Graphics2D g2d = (Graphics2D) strategy.getDrawGraphics();
 		g2d.setPaint(backgroundGradient);
-		g2d.fillRect(0, 0, (int) SIZE.width, SIZE.height);
+		// Fills the rectangle corresponding to the background
+		g2d.fillRect(0, 0, SIZE.width, SIZE.height);
 
-		// TODO Set this to draw only objects that are visible
+		// TODO renders each game object by calling the renderer
+		// TODO (Andy) Set this to draw only objects that are visible
 		for (GameEntity ent : ObjectManager.getObjects())
-			if ( ent != null && ent.renderer != null )
+			if (ent != null && ent.renderer != null)
 				ent.renderer.renderObject(g2d, alpha);
 
+		// Renders the GUI
 		renderGUI(g2d);
 
+		// Dispose of the old buffer strategy
 		g2d.dispose();
+		// Show the new strategy (ie swap the buffers)
 		strategy.show();
+		// Sync the buffers
 		Toolkit.getDefaultToolkit().sync();
 	}
 
+	/**
+	 * Renders the GUI
+	 * 
+	 * @param g2d
+	 *            the graphics context to render to
+	 */
 	public void renderGUI(Graphics2D g2d) {
 		int leftmargin = 12;
 		g2d.setColor(Color.green.brighter());
-		g2d.drawString("Objects on screen: " + ObjectManager.getObjects().size(), leftmargin, 20);
+		g2d.drawString("Objects on screen: "
+				+ ObjectManager.getObjects().size(), leftmargin, 20);
 		g2d.drawString(GameThread.getMTInfo(), leftmargin, 40);
 
 	}
